cleanups and readme.md
