--- src/main/java/org/basseur/adventofcode/advent2018/Application.java
@@ -1,17 +1,0 @@
-package org.basseur.adventofcode.advent2018;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * The Spring Boot Application.
- *
- * @author Michelle Fernandez Bieber
- */
-@SpringBootApplication
-public class Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-}
