--- src/main/java/org/basseur/adventofcode/advent2018/ProblemStatusEnum.java
@@ -1,17 +1,0 @@
-package org.basseur.adventofcode.advent2018;
-
-/**
- * Contains the possible statuses for each puzzle.
- *
- * @author Michelle Fernandez Bieber
- */
-public enum ProblemStatusEnum {
-    /** if the puzzle has been solved */
-    SOLVED,
-
-    /** if the puzzle is being worked on */
-    IN_PROGRESS,
-
-    /** if the puzzle hasn't been started yet */
-    UNSOLVED
-}
