--- src/main/java/org/basseur/adventofcode/advent2018/utils/FileReaders.java
@@ -1,17 +1,0 @@
-package org.basseur.adventofcode.advent2018.utils;
-
-import org.springframework.stereotype.Component;
-
-import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-import java.util.Scanner;
-
-/**
- * File readers to load and parse input files for the puzzles.
- */
-@Component
-public class FileReaders {
-
-}
