+++ src/main/java/org/haffson/adventofcode/Application.java
@@ -1,0 +1,17 @@
+package org.haffson.adventofcode;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * The Spring Boot Application.
+ *
+ * @author Michelle Fernandez Bieber
+ */
+@SpringBootApplication
+public class Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+}
