+++ src/main/java/org/haffson/adventofcode/ProblemStatusEnum.java
@@ -1,0 +1,17 @@
+package org.haffson.adventofcode;
+
+/**
+ * Contains the possible statuses for each puzzle.
+ *
+ * @author Michelle Fernandez Bieber
+ */
+public enum ProblemStatusEnum {
+    /** if the puzzle has been solved */
+    SOLVED,
+
+    /** if the puzzle is being worked on */
+    IN_PROGRESS,
+
+    /** if the puzzle hasn't been started yet */
+    UNSOLVED
+}
