+++ src/main/java/org/haffson/adventofcode/days/day01/Day01.java
@@ -1,0 +1,61 @@
+package org.haffson.adventofcode.days.day01;
+
+import org.haffson.adventofcode.ProblemStatusEnum;
+import org.haffson.adventofcode.days.Days;
+import org.haffson.adventofcode.utils.FileReaders;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import java.util.HashMap;
+
+/**
+ * Implementation for <i>Day 1: Chronal Calibration</i>.
+ */
+@Component
+public class Day01 implements Days {
+
+    /** The puzzle status {@code HashMap} */
+    private final HashMap<String, ProblemStatusEnum> problemStatus;
+
+    /**
+     * Causes the input file to be parsed into the frequencies array ({@code frequencies}).
+     *
+     * @param fileReaders {@code @Autowired} fileReader //TODO: inject what you need
+     */
+    @Autowired
+    Day01(FileReaders fileReaders) {
+        this.problemStatus = new HashMap<>();
+        this.problemStatus.put("1", ProblemStatusEnum.UNSOLVED);
+        this.problemStatus.put("2", ProblemStatusEnum.UNSOLVED);
+    }
+
+    @Override
+    public int getDay() {
+        return 1;
+    }
+
+    @Override
+    public HashMap<String, ProblemStatusEnum> getProblemStatus() {
+        return problemStatus;
+    }
+
+    @Override
+    public String firstPart() {
+        return "Part 1 - Frequency: " + calculateFrequency();
+    }
+
+    @Override
+    public String secondPart() {
+        return null;
+    }
+
+    /**
+     * Primary method for Day 1, Part 1.
+     * Calculates the final frequency as the sum of all frequencies.
+     *
+     * @return the final frequency
+     */
+    private int calculateFrequency() {
+        return 0;
+    }
+}
