--- src/test/java/org/basseur/adventofcode/advent2018/ArchUnitRulesFromPlantUML.java
@@ -1,29 +1,0 @@
-package org.basseur.adventofcode.advent2018;
-
-import com.tngtech.archunit.core.domain.JavaClasses;
-import com.tngtech.archunit.core.importer.ClassFileImporter;
-import com.tngtech.archunit.core.importer.ImportOption;
-import org.junit.Test;
-
-import java.net.URL;
-
-import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.classes;
-import static com.tngtech.archunit.library.plantuml.PlantUmlArchCondition.Configurations.consideringOnlyDependenciesInDiagram;
-import static com.tngtech.archunit.library.plantuml.PlantUmlArchCondition.adhereToPlantUmlDiagram;
-
-
-public class ArchUnitRulesFromPlantUML {
-
-    URL myDiagram = getClass().getClassLoader().getResource("adventOfCodeArchitecture.puml");
-
-    private final JavaClasses allClasses = new ClassFileImporter()
-            .withImportOption(new ImportOption.DontIncludeTests())
-            .importPackages("org.basseur.adventofcode");
-
-    @Test
-    public void classesShouldAdhereToPlantUmlDiagram() {
-        classes().should(adhereToPlantUmlDiagram(myDiagram, consideringOnlyDependenciesInDiagram()))
-                .check(allClasses);;
-    }
-
-}
