--- src/test/java/org/basseur/adventofcode/advent2018/ArchitectureUnitTest.java
@@ -1,4 +1,0 @@
-package org.basseur.adventofcode.advent2018;
-
-public interface ArchitectureUnitTest {
-}
