+++ src/test/java/org/haffson/adventofcode/ArchitectureUnitTest.java
@@ -1,0 +1,4 @@
+package org.haffson.adventofcode;
+
+public interface ArchitectureUnitTest {
+}
