+++ src/test/java/org/haffson/adventofcode/ArchitectureUnitTestSuite.java
@@ -1,0 +1,11 @@
+package org.haffson.adventofcode;
+
+import org.junit.experimental.categories.Categories;
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+
+@RunWith(Categories.class)
+@Categories.IncludeCategory(ArchitectureUnitTest.class)
+@Suite.SuiteClasses({ArchitectureUnitTesting.class})
+public class ArchitectureUnitTestSuite {
+}
