+++ src/test/java/org/haffson/adventofcode/utils/FileReadersTest.java
@@ -1,0 +1,18 @@
+package org.haffson.adventofcode.utils;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.test.context.junit4.SpringRunner;
+
+
+@RunWith(SpringRunner.class)
+public class FileReadersTest {
+
+    //TODO: Meaningful tests for readFileIntoArrayOfIntegers and readFileIntoStringList
+
+    //TODO: not easy to test, but we should!
+    @Test
+    public void openFileReturnScannerThrowsExceptionIfFileIsNotFound() {
+    }
+
+}
